Move loadgen out of internal
