+++ cmd/apmbench/bench_otlp.go
@@ -12,11 +12,11 @@ import (
 	"go.uber.org/zap"
 	"golang.org/x/time/rate"
 
-	loadgencfg "github.com/elastic/apm-perf/internal/loadgen/config"
 	"github.com/elastic/apm-perf/internal/telemetrygen/common"
 	"github.com/elastic/apm-perf/internal/telemetrygen/logs"
 	"github.com/elastic/apm-perf/internal/telemetrygen/metrics"
 	"github.com/elastic/apm-perf/internal/telemetrygen/traces"
+	loadgencfg "github.com/elastic/apm-perf/loadgen/config"
 )
 
 func commonConfigWithHTTPPath(httpPath string) common.Config {
