+++ loadgen/events/apm-go-minimal-2.0.0.ndjson
