+++ loadgen/events/apm-python-minimal-6.7.2.ndjson
