+++ loadgen/events/apm-ruby-minimal-4.5.0.ndjson
